[EDITED] Naming of project feature
