+++ Frontend/VIAProMa/Assets/Prefabs/Visualizations/Resources/StickyNote.prefab
