+++ Frontend/VIAProMa/Assets/Scripts/SaveLoadSystem/Serializers/StickyNoteSerializer.cs.meta
